# Optimised ExtractAllQuestIds ExtractAllItemIds