Based on 4 tests of 10000 entries:

New code metrics:
Time taken by function: 19 seconds
Time taken by function: 20 seconds
Time taken by function: 21 seconds
Time taken by function: 23 seconds

Old metric:
Time taken by function: 111 seconds
Time taken by function: 111 seconds
Time taken by function: 112 seconds
Time taken by function: 113 seconds