@zhangbo8418 2024-09-05
Anyway, whispers are always in English and will not be based on dbc.